+++ src/Shipment/Models/ShippingMethodProxy.php
@@ -1,0 +1,21 @@
+<?php
+
+declare(strict_types=1);
+
+/**
+ * Contains the ShippingMethodProxy class.
+ *
+ * @copyright   Copyright (c) 2023 Attila Fulop
+ * @author      Attila Fulop
+ * @license     MIT
+ * @since       2023-01-12
+ *
+ */
+
+namespace Vanilo\Shipment\Models;
+
+use Konekt\Concord\Proxies\ModelProxy;
+
+class ShippingMethodProxy extends ModelProxy
+{
+}